Properly reset region_highlight entry
If ZSH_AUTOSUGGEST_HIGHLIGHT_STYLE was something like `fg=7`, when
adding it to `region_highlight`, it would be replaced with `fg=white`
implicitly.
This means that the entry is not removed in
_zsh_autosuggest_highlight_reset and the suggestion highlighting
remains.

This uses the ZSH 5.9 `memo` feature which allows us to tag entries in
region_highlight and remove entry by tag instead of by exact contents.

Fixes #752
Ref: https://zsh.sourceforge.io/Doc/Release/Zsh-Line-Editor.html#Zle-Widgets